requirements.txt
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.13
PyYAML==5.3.1
Flask==1.1.2
pymysql==1.0.2
pytest==6.2.4
requests==2.25.1
boto3==1.17.49
botocore==1.20.49
pandas==1.0.3
s3fs==2021.5.0
scikit_learn==0.21.2
